adding files to 3.5.0

+++ app/assets/javascripts/surveyor/surveys.js.coffee
@@ -1,4 +1,4 @@
-# Copyright © 2011-2017 MUSC Foundation for Research Development~
+# Copyright © 2011-2019 MUSC Foundation for Research Development~
 # All rights reserved.~
 
 # Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:~
@@ -18,36 +18,47 @@
 # INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR~
 # TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.~
 $(document).ready ->
+  $("[data-toggle='tooltip']").tooltip()
+
   ### Survey Table ###
-  $(document).on 'click', '.edit-survey', ->
-    survey_id = $(this).data('survey-id')
+  $(document).on 'change', '.survey-actions', ->
+    $selected = $(this).find('option:selected')
+    $this = $(this)
 
-    $.ajax
-      type: 'get'
-      url: "/surveyor/surveys/#{survey_id}"
-
-  $(document).on 'click', '.delete-survey', ->
-    survey_id = $(this).data('survey-id')
-
-    swal {
-      title: I18n['swal']['swal_confirm']['title']
-      text: I18n['swal']['swal_confirm']['text']
-      type: 'warning'
-      showCancelButton: true
-      confirmButtonColor: '#DD6B55'
-      confirmButtonText: 'Delete'
-      closeOnConfirm: true
-    }, ->
+    if $selected.data('url')
       $.ajax
-        type: 'delete'
-        url: "/surveyor/surveys/#{survey_id}.js"
+        type: $selected.data('method') || 'get'
+        dataType: 'script'
+        url: $selected.data('url')
+        success: ->
+          $this.selectpicker('val', '')
+    else if $selected.hasClass('delete-survey')
+      survey_id = $selected.data('survey-id')
+      swal {
+        title: I18n['swal']['swal_confirm']['title']
+        text: I18n['swal']['swal_confirm']['text']
+        type: 'warning'
+        showCancelButton: true
+        confirmButtonColor: '#DD6B55'
+        confirmButtonText: 'Delete'
+        closeOnConfirm: true
+      }, ->
+        $.ajax
+          type: 'delete'
+          dataType: 'script'
+          url: "/surveyor/surveys/#{survey_id}"
+          success: ->
+            $this.selectpicker('val', '')
 
-  $(document).on 'click', '.preview-survey', ->
-    survey_id = $(this).data('survey-id')
+  $(document).on 'load-success.bs.table', '.survey-table, .form-table', ->
+    $('.selectpicker').selectpicker()
 
-    $.ajax
-      type: 'get'
-      url: "/surveyor/surveys/#{survey_id}/preview.js"
+  ### Survey Modal ###
+  $(document).on 'hide.bs.modal', '#modal_place', ->
+    if $(this).children("#survey-modal").length > 0
+      $('.survey-table').bootstrapTable('refresh')
+    else if $(this).children("#form-modal").length > 0
+      $('.form-table').bootstrapTable('refresh')
 
   $(document).on 'click', '.add-section', ->
     $.ajax
@@ -100,10 +111,10 @@ $(document).ready ->
   $(document).on 'change', '.select-depender, .select-question-type', ->
     send_update_request($(this), $(this).val())
 
-  $(document).on 'focusout', '#survey-modal input[type="text"], #survey-modal textarea', ->
+  $(document).on 'focusout', '#survey-modal input[type="text"], #form-modal input[type="text"]:not([id$="-surveyable"]), #survey-modal textarea, #form-modal textarea', ->
     send_update_request($(this), $(this).val())
 
-  $(document).on 'change', '#survey-modal input[type="checkbox"]', ->
+  $(document).on 'change', '#survey-modal input[type="checkbox"], #form-modal input[type="checkbox"]', ->
     send_update_request($(this), $(this).prop('checked'))
 
   $(document).on 'change', '.is-dependent', ->
